perf: optimize office filtering by precomputing tone and tag
- Moved expensive regex-based tone and tag classification from filter loops to the data loading phase in `useData.ts`.
- Updated `Office` interface and centralized `OfficeTone` and `OfficeTag` types in `src/types/index.ts`.
- Refactored `HomePage`, `ReviewPage`, `CountryPage`, and `CompanyPage` to use the precomputed properties.
- Achieved a ~35x speedup in the filtering logic (6ms vs 220ms for 10k iterations).
