+++ src/pages/CountryPage.tsx
@@ -5,7 +5,6 @@ import Photo from "../components/Photo";
 import Monogram from "../components/Monogram";
 import FlagChip from "../components/FlagChip";
 import MapView, { type MapFocus } from "../components/MapView";
-import { typeTag } from "../utils/typeTag";
 
 interface StatProps {
   n: number;
@@ -138,7 +137,7 @@ export default function CountryPage() {
                   </Link>
                   <div className="gof-crow-offices">
                     {list.map((o) => {
-                      const tag = typeTag(o.officeType);
+                      const tag = o.tag;
                       const isActive = activeId === o.id;
                       const isHover = hoverId === o.id;
                       return (
