+++ src/pages/HomePage.tsx
@@ -6,7 +6,7 @@ import MapView, { type MapFocus } from "../components/MapView";
 import Monogram from "../components/Monogram";
 import FlagChip from "../components/FlagChip";
 import Photo from "../components/Photo";
-import { REGION_ORDER, truncate, typeTag } from "../utils/typeTag";
+import { REGION_ORDER, truncate } from "../utils/typeTag";
 import { useData } from "../hooks/useData";
 
 type View = "grid" | "map";
@@ -84,7 +84,7 @@ export default function HomePage() {
       const co = companyById[o.companyId];
       if (!co) return false;
       if (region && o.region !== region) return false;
-      if (otype && typeTag(o.officeType).tone !== otype) return false;
+      if (otype && o.tone !== otype) return false;
       if (industry && co.industry !== industry) return false;
       if (needle) {
         const s = (
@@ -329,7 +329,7 @@ interface OfficeTileProps {
 }
 
 function OfficeTile({ office, company, onClose, onReadMore }: OfficeTileProps) {
-  const tag = typeTag(office.officeType);
+  const { tag } = office;
   const summary = truncate(company.description, 160);
   return (
     <div className="gof-mapcard" role="dialog" aria-label={`${company.name} — ${office.city}`}>
