+++ src/types/index.ts
@@ -32,6 +32,13 @@ export interface OfficeVerification {
   model?: string;
 }
 
+export type OfficeTone = "hq" | "reg" | "rnd";
+
+export interface OfficeTag {
+  short: string;
+  tone: OfficeTone;
+}
+
 export interface Office {
   id: string;
   companyId: string;
@@ -42,6 +49,8 @@ export interface Office {
   address: string;
   postalCode: string;
   officeType: string;
+  tone: OfficeTone;
+  tag: OfficeTag;
   latitude?: number;
   longitude?: number;
   contactUrl?: string;
